Feature/starting angular (#6)
* Simply ng new dutch-app

* Simply move dutch-app to ClientApp

* Tweak directory settings for our app

Couple differences from Shawn's video.  Additionally the index property
in angular.json seems to be required now.  The ng build worked, which
was the main goal with this commit.

* Really need to gitignore some of this stuff.

Added Angular scripts to shop and removed existing html.  Change the
selector for the component to the-shop, add this special tag to the shop
view.

* Tweak app component and html

* Build a productList component and include it